A lot of reorg and development of phase detection
